refactored some code to make some common services. Also should make implementing an actual backend much easier
