+++ managementTool/src/routes/+page.svelte
@@ -1,5 +1,6 @@
-<script>
+<script lang="ts">
     import Manager from '../components/Manager.svelte';
+    import type { EmployeeModel } from '../models/employee.model';
 </script>
 
 <h1 class="text-3xl font-bold underline">Management Helper Tool</h1>
