--- managementTool/src/services/mockData.service.svelte
@@ -1,3 +1,0 @@
-<script lang="ts">
-    
-</script>
